+++ Bastion/src/isaac/bastion/manager/BastionBlockManager.java
@@ -20,7 +20,8 @@ import org.bukkit.entity.Player;
 import isaac.bastion.Bastion;
 import isaac.bastion.BastionBlock;
 import isaac.bastion.BastionType;
-import isaac.bastion.events.BastionDamagedEvent;
+import isaac.bastion.event.BastionDamagedEvent;
+import isaac.bastion.event.BastionDamagedEvent.Cause;
 import isaac.bastion.storage.BastionBlockStorage;
 import vg.civcraft.mc.citadel.Citadel;
 import vg.civcraft.mc.citadel.reinforcement.PlayerReinforcement;
@@ -32,8 +33,6 @@ public class BastionBlockManager {
 	
 	private HashMap<UUID, HashMap<String, Long>> cooldowns = new HashMap<UUID, HashMap<String, Long>>();
 	private BastionBlockStorage storage;
-
-	public enum Cause {PEARL, BLOCK, ELYTRA}
 	
 	public BastionBlockManager() {
 		storage = Bastion.getBastionStorage();
@@ -41,7 +40,7 @@ public class BastionBlockManager {
 	
 	//For block places
 	public void erodeFromPlace(Player player, Set<BastionBlock> blocking) {
-		erodeFromAction(player, blocking, Cause.BLOCK);
+		erodeFromAction(player, blocking, Cause.BLOCK_PLACED);
 	}
 
 	//For pearls
@@ -68,7 +67,7 @@ public class BastionBlockManager {
 			Set<BastionBlock> bastions = typeMap.get(key);
 			if(type.getBlocksToErode() < 0) {
 				for(BastionBlock bastion : bastions) {
-					double damage = cause == Cause.BLOCK ? bastion.getErosionFromBlock() : bastion.getErosionFromPearl();
+					double damage = cause == Cause.BLOCK_PLACED ? bastion.getErosionFromBlock() : bastion.getErosionFromPearl();
 					BastionDamagedEvent event = new BastionDamagedEvent(bastion, player, cause, damage);
 					Bukkit.getPluginManager().callEvent(event);
 					if(event.isCancelled()) continue;
@@ -79,7 +78,7 @@ public class BastionBlockManager {
 				for(int i = 0; i < ordered.size() && i < type.getBlocksToErode(); i++) {
 					int erode = rng.nextInt(ordered.size());
 					BastionBlock bastion = ordered.get(erode);
-					double damage = cause == Cause.BLOCK ? bastion.getErosionFromBlock() : bastion.getErosionFromPearl();
+					double damage = cause == Cause.BLOCK_PLACED ? bastion.getErosionFromBlock() : bastion.getErosionFromPearl();
 					BastionDamagedEvent event = new BastionDamagedEvent(bastion, player, cause, damage);
 					Bukkit.getPluginManager().callEvent(event);
 					if(event.isCancelled()) continue;
